Q{AbstractAnimation,Unified}Timer: Port runningAnimationCount() to qsizetype
Applying headersclean_check to private headers has pointed out that
QAbstractAnimationTimer::runningAnimationCount() narrows qsizetype to
int in its return statement.

Since this is private API, we can change the return type of the
functions called runningAnimationCount() (incl. the one in
QUnifiedTimer) from int to qsizetype, but have to adapt the users in
other modules first.

QtDeclarative is the only such module.

As a drive-by, make the functions const. This means we need a macro to
propagate the information of which signature to use to qtdeclarative
(if it was just the return type, we could use decltype() introspection
to automatically adjust the overriding function's return type).

Task-number: QTBUG-103530
Task-number: QTBUG-126219
Change-Id: I1bb6b75f1539b453982d305f95b464d732aa5222
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>