QLoggingCategory: Deprecate non-static, non-forward-declared categories
Qt-internal logging categories should either be static or declared in a
header and therefore shared between multiple translation units. This way
we minimize the number of exposed symbols. We want to minimize the
number of exposed symbols in order to reduce the chance of name clashes.

In the same vein, when exporting a Qt-internal logging category it
should be declared explicitly as Qt-internal. The code importing it may
not be a Qt library after all.

Task-number: QTBUG-67692
Change-Id: If1c045471fe52226a34cc94a5fe9b0c9a2de56b5
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>