Add support for VRS
The basic Metal plumbing for letting an MTLRasterizationRateMap through
is important for Vision Pro, that seems to rely heavily on VRS to reduce
the GPU load via eye-tracked, dynamic foveation.

Metal's approach is kind of different than everyone else. For a more
complete, forward looking story, introduce support for some of the D3D12
and Vulkan VRS features. (these seem to be identical feature-wise, with
Vulkan essentially copying D3D, just with a lot more convoluted API) A
certain fragmentation is however inevitable.

This way we support:

- setting a per-draw rate (say I want to draw this mesh with 4x4 rate,
  i.e. 4x reduced resolution both horizontally and vertically). This is
  supported with D3D12/Vulkan, and could also be supported with the
  Qualcomm OpenGL ES extension, although that's not currently
  implemented.

- setting a texture-based rate map, D3D12/Vulkan style. This implies
  starting to support R8_UINT textures since R8_UNORM, while the
  difference isn't necessarily relevant here, will be rejected by some
  APIs. This is done by setting a QRhiTexture on a QRhiShadingRateMap (a
  new lightweight, wrapper-style object that does not do much on its
  own, just contains either a texture or some other native object)

- setting an MTLRasterizationRateMap on a QRhiShadingRateMap. This is
  only supported on Metal, naturally, and is realized by an overload of
  createFrom(). Conversely, setting a QRhiTexture (or attempting to set
  a per-draw rate) will not work with Metal.

We do not offer ways to create a rate map directly via Qt, due to the
underlying API differences. So for the user this involves either
uploading a Grayscale8 or similar QImage into an R8UI QRhiTexture, or by
setting up an MTLRasterizationRateMap directly with Metal and passing
that in to the QRhiShadingRateMap instead of a QRhiTexture.

We also do not care about the bizarre 'render to a texture, get a buffer
with scale rates, and then do a scaling pass using the buffer and MSL
intrinsics in the shader' approach of Metal. For now we assume that in
case of the Vision Pro this is taken care of by the XR compositor, and
is not something the applications (Qt) implements.

Some D3D12/Vulkan VRS features are left as potential future
enhancements. For example, when it comes to combining the two rates, we
always do MAX, there is no API to control the combiners.

The third mode of D3D12/Vulkan VRS, per-triangle (writing to a special
variable in the vertex shader) is not planned to be supported now or in
the future.

The Vulkan backend gets quite a lot of changes, mainly due to the
difficulty of dealing with extensions in a portable and robust manner in
that API, and the API versioning design mistakes in Vulkan 1.1 and
beyond. Now both swapchain and texture render pass objects go through
vkCreateRenderPass2, when supported, but the KHR variant still, not the
1.2+ core functions, in order to play nice with Vulkan 1.1 and the
Android ecosystem.

Some QRhi docs are enhanced, e.g. around QRhiRenderPassDescriptor, since
as the manual test demonstrates, dynamically taking VRS into use in a
scene involves certain steps (new/compatible QRhiRenderPassDescriptor is
needed, rebuilding the render target) that are not trivial. (these are
just the usual consequences of Vulkan's VkRenderPass permeating
everything in the rendering engines)

The manual test can be used to exercise both per-draw and image-based
VRS with D3D12 and Vulkan. It also includes a limited example of setting
up a rate map with Metal (but the test app does not implement the
special scaling) A number of things, such as operating with MSAA
enabled or using VRS combined with multiview are not currently
exercised by the test app, and are left as a future exercise to
verify.

Task-number: QTBUG-126297
Change-Id: I5210f4cff6b8360b6bb47cdbe8d9caee1c29b8a5
Reviewed-by: Andy Nichols <andy.nichols@qt.io>