QLibrary: move prefixes_sys() to the header and make it prefix_sys()
There's only one prefix for Unix ("lib") and none for Windows, so
there's no need to allocate a QList for it in this function. For one of
the two callers, it was also possible to avoid the QList and QString
allocations, at the expense of testing the same file twice on Windows.
We can't do the same for Unix systems because we may add more elements
to the QList.

Change-Id: Ic0adfa808d28487a8303fffd17d9e6d35aaa79f4
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>